Lesson1_task3_finish Внес исправления: 1) Отформатировал текст через Ctrl+Alt+L 2) Изменил seconds с типа Int на тип Short

+++ src/main/kotlin/lesson1/lesson1_task4.kt
@@ -1,19 +1,21 @@
 package org.example.lesson1
 
-fun main(){
-    val length:Long = 40_868_600_000L
-    val age:Byte = 27
-    val partOfDay:Float = 0.075f
-    val seconds:Int = 6480
-    val partOfYear:Double = 0.00020547945205479453
-    val apogee:Int = 327000
-    print("""
+fun main() {
+    val length: Long = 40_868_600_000L
+    val age: Byte = 27
+    val partOfDay: Float = 0.075f
+    val seconds: Short = 6480
+    val partOfYear: Double = 0.00020547945205479453
+    val apogee: Int = 327000
+    print(
+        """
         $length
         $age
         $partOfDay
         $seconds
         $partOfYear
         $apogee
-    """.trimIndent())
+    """.trimIndent()
+    )
 }
